Am now able to open tabs! when given an ID as a command line argument, however:
- Currently I cannot pass an optional interger argument that would represent
a site's ID number. Currently you are required to put a 0 as the ID number in
order to perform no action.
I would like to be able to leave an argument blank, but if the argument is there
it will be an interger of the ID of the article to open.