Lots has happened!
I decided that the best way to rank articles was to do that in a seperate script rather than wait on input from a user in the renewecon_read script.
Hence I created the renewecon_rank script.
The script takes an article ID and a ranks as command line variable and updates the url_database to the rank.
It also performs some basic checks to ensure that the article's ranks is not too silly and that the argument exists.

TODO:
Still need to comment all over the script before I get too carried away.